+++ lib/src/comm.cpp
@@ -110,12 +110,12 @@ sock connect_to_camera(int port) {
   return 0;
 }
 
-uint32_t to_fuji_size_prefix(uint32_t sizeBytes) {
+size_t to_fuji_size_prefix(size_t sizeBytes) {
   // TODO, 0x endianess
   return sizeBytes;
 }
 
-uint32_t from_fuji_size_prefix(uint32_t sizeBytes) {
+size_t from_fuji_size_prefix(size_t sizeBytes) {
   // TODO, 0x endianess
   return sizeBytes;
 }
@@ -153,14 +153,14 @@ void receive_data(native_socket sockfd, void* data, size_t sizeBytes) {
   }
 }
 
-void fuji_send(native_socket sockfd, void const* data, uint32_t sizeBytes) {
-  uint32_t const size = to_fuji_size_prefix(sizeBytes + sizeof(uint32_t));
+void fuji_send(native_socket sockfd, void const* data, size_t sizeBytes) {
+  size_t const size = to_fuji_size_prefix(sizeBytes + sizeof(uint32_t));
   send_data(sockfd, &size, sizeof(uint32_t));
   send_data(sockfd, data, sizeBytes);
 }
 
-size_t fuji_receive(native_socket sockfd, void* data, uint32_t sizeBytes) {
-  uint32_t size = 0;
+size_t fuji_receive(native_socket sockfd, void* data, size_t sizeBytes) {
+  size_t size = 0;
   receive_data(sockfd, &size, sizeof(size));
   size = from_fuji_size_prefix(size);
   if (size < sizeof(size)) {
